backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==5.0.1
attrs==24.2.0
azure-functions==1.21.3
blis==0.7.11
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==7.1.2
cloudpathlib==0.20.0
confection==0.1.5
cryptography==43.0.3
cymem==2.0.10
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0.tar.gz#sha256=0673b3ce4a28a19e99dc0fc3f8f7fe67b936c41749e6b70f778ec5f9f78f3a66
Flask==2.0.3
frozenlist==1.5.0
fuzzywuzzy==0.18.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
murmurhash==1.0.11
numpy==2.0.2
openai==0.28.1
packaging==24.2
pathlib_abc==0.1.1
pathy==0.11.0
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
preshed==3.0.9
propcache==0.2.0
pycparser==2.22
pydantic==1.8.2
pydantic_core==2.27.1
Pygments==2.18.0
pypdfium2==4.30.0
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
smart-open==6.4.0
spacy==3.0.9
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.0.17
tqdm==4.67.0
typer==0.3.2
typing_extensions==4.12.2
urllib3==2.2.3
wasabi==0.10.1
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.0
yarl==1.18.0
